pyproject.toml
[project]
name = "rt-sch-cli-equity-pipeline"
version = "0.1.0"
description = "Real-Time School Climate & Equity Pipeline using Kafka, Databricks, dbt, Snowflake, and Power BI"
authors = [
    { name="Dylan Picart", email="dylanpicart@mail.adelphi.edu" }
]
requires-python = ">=3.10"

# Core universal dependencies
dependencies = [
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "pandas>=2.2.3",
    "google-cloud-storage>=2.17.0",
    "google-auth>=2.30.0"
]

[project.optional-dependencies]

# Kafka streaming components
kafka = [
    "confluent-kafka>=2.4.0",
]

# Databricks / Spark components
databricks = [
    "pyspark>=3.5.0",
]

# Snowflake ingestion + modeling
snowflake = [
    "snowflake-connector-python>=3.7.0",
    "snowflake-sqlalchemy>=1.5.0",
]

# Developer tools
dev = [
    "black>=24.3.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
# Disable automatic package discovery – we’re using this project
# primarily as a “mono-repo” with scripts and configs.
packages = []

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
